backend/requirements.txt
asgiref>=3.7
diff-match-patch==20230430
distlib==0.3.8

filelock==3.15.4
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
setuptools<81
sqlparse>=0.5
tablib==3.5.0
tzdata
virtualenv==20.26.2
whitenoise==6.6.0


# Django y REST Framework
Django>=4.2,<5.0
django-filter>=23.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-import-export==4.0.6
django-js-asset==2.2.0
django-model-utils==4.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt>=5.3.2

# Autenticación y Seguridad
django-axes<7.0
django-ratelimit>=4.1.0

# Base de datos
psycopg2-binary>=2.9.0

# Validación y utilidades
phonenumbers>=8.13.0
python-dateutil>=2.8.0
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"

# Email y templates
django-templated-mail>=1.1.1

# Documentación API
drf-spectacular>=0.27.0

# PDF Generation
weasyprint>=60.0
reportlab>=4.0.0
qrcode[pil]>=7.4.2

# Export XLSX
openpyxl>=3.1.0

# Tareas asíncronas
celery>=5.3.0
redis>=5.0.0

# Backups
django-dbbackup<5.0

# Recurrencias calendario
icalendar>=5.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Desarrollo
django-debug-toolbar>=4.2.0
ipython>=8.12.0

# Producción
gunicorn>=21.2.0
python-decouple>=3.8
whitenoise>=6.5.0

# Monitoreo (opcional)
# sentry-sdk>=1.32.0
